SP2-157 - adding api for heating grid and seperating images controller
